feature: merging beacon and contribute and creating two separate binaries
