+++ mpc/package.json
@@ -21,7 +21,7 @@
   "dependencies": {
     "aws-sdk": "^2.1259.0",
     "axios": "^1.1.3",
-    "chai": "^4.3.6",
+    "chai": "^4.3.7",
     "chai-as-promised": "^7.1.1",
     "chalk": "^4.1.2",
     "commander": "^9.4.1",
@@ -33,7 +33,7 @@
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-prettier": "^4.0.0",
     "esm": "^3.2.25",
-    "mocha": "^10.0.0",
+    "mocha": "^10.1.0",
     "prettier": "^2.6.2",
     "promise": "^8.3.0",
     "promptly": "^3.2.0",
